Move GetTestClient from vmutils to testutils so it can be used in other SDK packages
